fix(build): resolve Windows MSVC warnings and modernize build pipeline with Ninja
- Fix C4129 unrecognized escape characters in QTBundle SYSTEM_PYTHONPATH/HOME by normalizing to forward slashes and using os.pathsep / platform SEP
- Fix C4805 unsafe mix of bool and int in PyRootNode plugin validation
- Fix C4477 sprintf format mismatch by replacing platform PAD %06d with standard %06zu
- Fix C4311/C4312 pointer truncation in FileMMap (HANDLE stored in long -> intptr_t) and File.cpp (ShellExecute -> INT_PTR)
- Fix C4834 [[nodiscard]] return value warnings on QFile::open in RvNetworkDialog and StreamConnection
- Fix C4312 pointer cast warnings on 0xdeadc0de sentinels via uintptr_t across FrameBuffer, IPImage, GLVideoDevice, and thread_group
- Fix C4477 format string mismatch with pthread_t in thread_group via thread_to_ptr helper
- Update src/pub submodule with UTF-8 re-encoding to eliminate C4828 codepage warnings
- Modernize Windows CI with Ninja generator, -Z7 debug format, and sccache compilation caching

+++ src/lib/base/TwkUtil/FileMMap.cpp
@@ -105,8 +105,8 @@ namespace TwkUtil
 
         OFSTRUCT of;
 
-        fileHandle = (long)CreateFile(filename.c_str(), GENERIC_READ, FILE_SHARE_READ, NULL, OPEN_EXISTING,
-                                      FILE_FLAG_SEQUENTIAL_SCAN | FILE_ATTRIBUTE_NORMAL, NULL);
+        fileHandle = (intptr_t)CreateFile(filename.c_str(), GENERIC_READ, FILE_SHARE_READ, NULL, OPEN_EXISTING,
+                                          FILE_FLAG_SEQUENTIAL_SCAN | FILE_ATTRIBUTE_NORMAL, NULL);
 
         if (((HANDLE)fileHandle) == INVALID_HANDLE_VALUE)
         {
@@ -115,7 +115,7 @@ namespace TwkUtil
 
         if (reallyMMap)
         {
-            mapHandle = (long)CreateFileMapping(((HANDLE)fileHandle), NULL, PAGE_READONLY, 0, 0, NULL);
+            mapHandle = (intptr_t)CreateFileMapping(((HANDLE)fileHandle), NULL, PAGE_READONLY, 0, 0, NULL);
 
             if (!mapHandle || ((HANDLE)mapHandle) == INVALID_HANDLE_VALUE)
             {
